requirements.txt
grpcio==1.30.0
grpcio-tools==1.30.0
numpy==1.20.3
torch==1.8.1+cu111
pokerenv